+++ packages/flutter/lib/material.dart
@@ -48,12 +48,8 @@ export 'src/material/floating_action_button.dart';
 export 'src/material/flutter_logo.dart';
 export 'src/material/grid_tile.dart';
 export 'src/material/grid_tile_bar.dart';
-export 'src/material/icon.dart';
 export 'src/material/icon_button.dart';
-export 'src/material/icon_theme.dart';
-export 'src/material/icon_theme_data.dart';
 export 'src/material/icons.dart';
-export 'src/material/image_icon.dart';
 export 'src/material/ink_highlight.dart';
 export 'src/material/ink_splash.dart';
 export 'src/material/ink_well.dart';
