+++ packages/flutter/lib/src/material/about.dart
@@ -12,9 +12,6 @@ import 'app_bar.dart';
 import 'debug.dart';
 import 'dialog.dart';
 import 'flat_button.dart';
-import 'icon.dart';
-import 'icon_theme.dart';
-import 'icon_theme_data.dart';
 import 'list_tile.dart';
 import 'page.dart';
 import 'progress_indicator.dart';
