+++ packages/flutter/lib/src/material/app_bar.dart
@@ -12,10 +12,7 @@ import 'package:flutter/widgets.dart';
 import 'back_button.dart';
 import 'constants.dart';
 import 'flexible_space_bar.dart';
-import 'icon.dart';
 import 'icon_button.dart';
-import 'icon_theme.dart';
-import 'icon_theme_data.dart';
 import 'icons.dart';
 import 'material.dart';
 import 'page.dart';
