+++ packages/flutter/lib/src/material/back_button.dart
@@ -4,7 +4,6 @@
 
 import 'package:flutter/widgets.dart';
 
-import 'icon.dart';
 import 'icon_button.dart';
 import 'icons.dart';
 import 'theme.dart';
