+++ packages/flutter/lib/src/material/data_table.dart
@@ -12,9 +12,6 @@ import 'checkbox.dart';
 import 'colors.dart';
 import 'debug.dart';
 import 'dropdown.dart';
-import 'icon.dart';
-import 'icon_theme.dart';
-import 'icon_theme_data.dart';
 import 'icons.dart';
 import 'ink_well.dart';
 import 'material.dart';
