+++ packages/flutter/lib/src/material/date_picker.dart
@@ -18,7 +18,6 @@ import 'colors.dart';
 import 'debug.dart';
 import 'dialog.dart';
 import 'flat_button.dart';
-import 'icon.dart';
 import 'icon_button.dart';
 import 'icons.dart';
 import 'ink_well.dart';
