+++ packages/flutter/lib/src/material/expand_icon.dart
@@ -7,7 +7,6 @@ import 'package:flutter/widgets.dart';
 
 import 'colors.dart';
 import 'debug.dart';
-import 'icon.dart';
 import 'icon_button.dart';
 import 'icons.dart';
 import 'theme.dart';
