+++ packages/flutter/lib/src/material/flutter_logo.dart
@@ -5,8 +5,6 @@
 import 'package:flutter/widgets.dart';
 
 import 'colors.dart';
-import 'icon_theme.dart';
-import 'icon_theme_data.dart';
 
 /// The Flutter logo, in widget form. This widget respects the [IconTheme].
 ///
