+++ packages/flutter/lib/src/material/paginated_data_table.dart
@@ -14,10 +14,7 @@ import 'card.dart';
 import 'data_table.dart';
 import 'data_table_source.dart';
 import 'dropdown.dart';
-import 'icon.dart';
 import 'icon_button.dart';
-import 'icon_theme.dart';
-import 'icon_theme_data.dart';
 import 'icons.dart';
 import 'progress_indicator.dart';
 import 'theme.dart';
