+++ packages/flutter/lib/src/material/tabs.dart
@@ -13,8 +13,6 @@ import 'app_bar.dart';
 import 'colors.dart';
 import 'constants.dart';
 import 'debug.dart';
-import 'icon_theme.dart';
-import 'icon_theme_data.dart';
 import 'ink_well.dart';
 import 'material.dart';
 import 'tab_controller.dart';
