+++ packages/flutter/lib/src/material/two_level_list.dart
@@ -6,9 +6,6 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
 import 'colors.dart';
-import 'icon.dart';
-import 'icon_theme.dart';
-import 'icon_theme_data.dart';
 import 'icons.dart';
 import 'list_tile.dart';
 import 'theme.dart';
