+++ packages/flutter/lib/widgets.dart
@@ -34,7 +34,12 @@ export 'src/widgets/framework.dart';
 export 'src/widgets/gesture_detector.dart';
 export 'src/widgets/grid_paper.dart';
 export 'src/widgets/heroes.dart';
+export 'src/widgets/icon.dart';
+export 'src/widgets/icon_data.dart';
+export 'src/widgets/icon_theme.dart';
+export 'src/widgets/icon_theme_data.dart';
 export 'src/widgets/image.dart';
+export 'src/widgets/image_icon.dart';
 export 'src/widgets/implicit_animations.dart';
 export 'src/widgets/layout_builder.dart';
 export 'src/widgets/locale_query.dart';
